requirements.txt
flet==0.7.4
python-dateutil==2.8.2